clases/2025-02/15_Airflow/airflow_class/requirements.txt
apache-airflow==2.11.0
apache-airflow-providers-fab
scikit-learn
pandas
numpy
lightgbm
joblib